requirements.txt
speedtest-cli
colorama
requests